[Tests] TestDeleteDocument now passes.
Completed the port of all changes from the first sharpen snapshot
to the second. Fixed a few issues related to refactoring.

+++ Couchbase.Lite/Couchbase.Lite/Attachment.cs
@@ -79,7 +79,7 @@ namespace Couchbase.Lite {
                     {
                         // Copy attachment body into the database's blob store:
                         var writer = BlobStoreWriterForBody(body, database);
-                        metadataMutable["length"] = writer.GetLength();
+                        metadataMutable["length"] = (long)writer.GetLength();
                         metadataMutable["digest"] = writer.MD5DigestString();
                         metadataMutable["follows"] = true;
                         database.RememberAttachmentWriter(writer);
